+++ src/Pim/Bundle/CatalogBundle/AttributeType/BooleanType.php
@@ -25,7 +25,7 @@ class BooleanType extends FlexBooleanType
                 'fieldType' => 'switch'
             ),
             array(
-                'name'      => 'translatable',
+                'name'      => 'localizable',
                 'fieldType' => 'switch',
                 'options'   => array(
                     'disabled'  => (bool) $attribute->getId(),