+++ src/Pim/Bundle/CatalogBundle/AttributeType/ImageType.php
@@ -39,7 +39,7 @@ class ImageType extends FlexImageType
                 )
             ),
             array(
-                'name'      => 'translatable',
+                'name'      => 'localizable',
                 'fieldType' => 'switch',
                 'options'   => array(
                     'disabled'  => (bool) $attribute->getId(),