+++ src/Pim/Bundle/CatalogBundle/AttributeType/MetricType.php
@@ -76,7 +76,7 @@ class MetricType extends FlexMetricType
                 'fieldType' => 'switch'
             ),
             array(
-                'name'      => 'translatable',
+                'name'      => 'localizable',
                 'fieldType' => 'switch',
                 'options'   => array(
                     'disabled'  => (bool) $attribute->getId(),