+++ src/Pim/Bundle/CatalogBundle/Form/View/ProductFormView.php
@@ -99,7 +99,7 @@ class ProductFormView
             $classes['scopable'] = true;
         }
 
-        if ($attribute->isTranslatable()) {
+        if ($attribute->isLocalizable()) {
             $classes['translatable'] = true;
         }